+++ app/src/main/kotlin/com/arturogutierrez/openticator/view/fragment/BaseFragmentExt.kt
@@ -1,11 +1,10 @@
 package com.arturogutierrez.openticator.view.fragment
 
 import android.support.annotation.StringRes
-import android.support.design.widget.BaseTransientBottomBar.Duration
 import android.support.design.widget.Snackbar
 import com.arturogutierrez.openticator.view.activity.BaseActivity
 
-fun BaseFragment.makeSnackbar(@StringRes resId: Int, @Duration duration: Int) {
+fun BaseFragment.makeSnackbar(@StringRes resId: Int, duration: Int) {
   val coordinatorLayout = (activity as BaseActivity).coordinatorLayout
   Snackbar.make(coordinatorLayout, resId, duration).show()
 }