+++ domain/src/main/kotlin/com/arturogutierrez/openticator/domain/cloud/SessionManager.kt
@@ -1,0 +1,11 @@
+package com.arturogutierrez.openticator.domain.cloud
+
+import io.reactivex.Flowable
+
+interface SessionManager {
+
+  val currentSession : Session
+
+  fun onSession() : Flowable<Session>
+
+}